treasury/hermes/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hermes-token-screener"
version = "10.0.0"
description = "Unified DeFi platform: token screening, smart money tracking, and multi-chain trading execution"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "TerexitariusStomp" },
]
keywords = ["defi", "token", "screener", "crypto", "solana", "evm", "smart-money", "dex", "trading", "multi-chain", "web3"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "requests>=2.28",
    "httpx[socks]>=0.27",
    "pysocks>=1.7",
    "python-dotenv>=1.0",
    "structlog>=24.0",
    "prometheus-client>=0.20",
    "rich>=13.0",
]

[project.optional-dependencies]
telegram = ["telethon>=1.30"]
dashboard = ["fastapi>=0.110", "uvicorn>=0.29"]
trading = [
    "web3>=6.0",
    "eth-account>=0.10",
    "solders>=0.21",
    "solana>=0.30",
    "PyYAML>=6.0",
    "py-clob-client>=0.25.0",
    "py-clob-client-v2>=1.0.0",
    "pmxt>=2.31.1",
]
ml = [
    "torch>=2.0",
    "transformers>=4.30",
    "bitsandbytes>=0.41",
    "numpy>=1.24",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
    "black>=24.0",
    "mypy>=1.10",
    "bandit>=1.7.9",
    "pip-audit>=2.7.3",
]
all = ["hermes-token-screener[telegram,dashboard,trading,ml,dev,worker]"]
worker = ["fastapi>=0.110", "uvicorn[standard]>=0.29"]

[project.urls]
Homepage = "https://github.com/TerexitariusStomp/hermes-token-screener"
Repository = "https://github.com/TerexitariusStomp/hermes-token-screener"

[project.scripts]
hermes-screener = "hermes_screener.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["hermes_screener"]

[tool.black]
target-version = ["py312"]
line-length = 120


[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501", "E722", "N806", "B905", "SIM105", "SIM108", "SIM115"]

[tool.ruff.lint.per-file-ignores]
"hermes_screener/trading/*.py" = ["F401", "F811", "F841", "E402", "B033"]
"hermes_screener/agents/*.py" = ["UP035", "SIM103", "F841"]
"hermes_screener/memory/*.py" = ["F401", "F841"]
"hermes_screener/models/*.py" = ["F401"]
"tests/*.py" = ["SIM102", "B011", "N805", "N802", "N803"]
"scripts/*.py" = [
    "UP006", "UP035", "UP045", "I001", "F541", "F841", "F401",
    "SIM102", "SIM113", "SIM114", "N803", "N805", "N802",
    "E402", "E741", "B007", "B033", "W293", "W291",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disable_error_code = ["prop-decorator", "no-any-return", "var-annotated", "assignment", "union-attr", "arg-type", "return-value", "import-untyped", "index", "attr-defined", "dict-item", "no-redef", "operator"]
